Upgrade to latest JUnit Foundation; fix unit test packaging * Upgrade to JUnit Foundation 9.1.2 * Remove redundant 'junit' from unit test package name * Expected failures should be reported safely
