+++ build.gradle
@@ -20,7 +20,7 @@ repositories {
 dependencies {
     compile 'com.epam.reportportal:client-java:4.0.12'
     compile 'com.epam.reportportal:commons-model:4.2.0'
-    compile 'com.nordstrom.tools:junit-foundation:9.1.1'
+    compile 'com.nordstrom.tools:junit-foundation:9.1.2'
     compile 'com.google.inject:guice:4.1.0'
 }
 ext {
@@ -32,4 +32,5 @@ test.doFirst {
 test {
 //  debug true
     testLogging.showStandardStreams = true
+    exclude 'com/epam/reportportal/junit/ExpectToFail.class'
 }
