+++ src/test/java/com/epam/reportportal/junit/ExpectToFail.java
@@ -1,0 +1,23 @@
+package com.epam.reportportal.junit;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class ExpectToFail {
+
+    @Test
+    public void testFailure() {
+        Assert.assertEquals(2, 1);
+    }
+    
+    @Test
+    public void testFailureWithCustomMessage() {
+        Assert.assertEquals("Failure msg", 2, 1);
+    }
+
+    @Test(expected = AssertionError.class)
+    public void expectedFailureAbsent() {
+        Assert.assertEquals(1, 1);
+    }
+    
+}
