+++ src/test/java/com/epam/reportportal/junit/SuiteOfSuiteTest.java
@@ -1,4 +1,4 @@
-package com.epam.reportportal.junit.junit;
+package com.epam.reportportal.junit;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
